Remove DesktopBuildClient
