+++ src/Compilers/Server/VBCSCompiler/VBCSCompiler.csproj
@@ -26,7 +26,6 @@
     <Compile Include="..\..\Shared\BuildServerConnection.cs" />
     <Compile Include="..\..\Shared\CoreClrAnalyzerAssemblyLoader.cs" />
     <Compile Include="..\..\Shared\DesktopAnalyzerAssemblyLoader.cs" />
-    <Compile Include="..\..\Shared\DesktopBuildClient.cs" />
     <Compile Include="..\..\Shared\ExitingTraceListener.cs" />
     <Compile Include="..\..\Shared\ShadowCopyAnalyzerAssemblyLoader.cs" />
   </ItemGroup>
