+++ src/Compilers/VisualBasic/vbc/Program.cs
@@ -29,7 +29,7 @@ namespace Microsoft.CodeAnalysis.VisualBasic.CommandLine
             ExitingTraceListener.Install();
 #endif
 
-            return DesktopBuildClient.Run(args, RequestLanguage.VisualBasicCompile, Vbc.Run);
+            return BuildClient.Run(args, RequestLanguage.VisualBasicCompile, Vbc.Run);
         }
 
         public static int Run(string[] args, string clientDir, string workingDir, string sdkDir, string tempDir, TextWriter textWriter, IAnalyzerAssemblyLoader analyzerLoader)
