+++ src/Compilers/VisualBasic/vbc/vbc.csproj
@@ -24,7 +24,6 @@
     <Compile Include="..\..\Shared\BuildClient.cs" />
     <Compile Include="..\..\Shared\BuildServerConnection.cs" />
     <Compile Include="..\..\Shared\CoreClrAnalyzerAssemblyLoader.cs" />
-    <Compile Include="..\..\Shared\DesktopBuildClient.cs" />
     <Compile Include="..\..\Shared\ExitingTraceListener.cs" />
     <Compile Include="..\..\Shared\DesktopAnalyzerAssemblyLoader.cs" />
     <Compile Include="..\..\Shared\Vbc.cs" />
